requirements.txt
streamlit>=1.30.0
markdown>=3.5
get-code-from-markdown>=1.0.0
defusedxml>=0.7.1
requests>=2.32.0
pypdf>=5.0.0
langchain>=0.2.0
langchain-community>=0.2.0
unstructured>=0.10.0,<0.18.0
python-pptx>=1.0.0
pyshorteners>=1.0.1
python-dotenv>=1.0.0
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-aiplatform>=1.40.0
firebase-admin>=6.4.0 